Save Deleted properties to API

+++ src/App.js
@@ -57,7 +57,11 @@ class App extends Component {
   render() {
 
     let notes = this.state.notes.map((val, key) => {
-      return <Note key={key} note={val} delete={() => this.deleteNote(key)} />
+
+      if (!val.deleted){ //Somente retorna Notes que nao foram deletados
+        return <Note key={key} note={val} delete={() => this.deleteNote(key)} />
+      }
+      
     })
 
     return (
